add; styles for home page
