+++ lib/pages/home/widgets/global_action_widget.dart
@@ -1,9 +1,15 @@
 import 'package:flutter/material.dart';
+import 'package:pet_project/common/theme/text_theme.dart';
 
 class GlobalActionWidget extends StatelessWidget {
   const GlobalActionWidget({
     Key? key,
+    required this.actionName,
+    required this.actionImage,
   }) : super(key: key);
+
+  final String actionName;
+  final AssetImage actionImage;
 
   @override
   Widget build(BuildContext context) {
@@ -12,17 +18,35 @@ class GlobalActionWidget extends StatelessWidget {
       child: Container(
         width: 64,
         height: 64,
-        decoration: const BoxDecoration(
+        decoration: BoxDecoration(
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black26.withOpacity(0.2),
+              spreadRadius: 1,
+              blurRadius: 5,
+              offset: const Offset(0, 2),
+            )
+          ],
           color: Colors.white,
-          borderRadius: BorderRadius.all(
+          borderRadius: const BorderRadius.all(
             Radius.circular(8),
           ),
         ),
         child: Column(
           mainAxisAlignment: MainAxisAlignment.spaceAround,
-          children: const [
-            Icon(Icons.pets),
-            Text('Dog'),
+          children: [
+            Image(
+              image: actionImage,
+              height: 24,
+              width: 24,
+            ),
+            Text(
+              actionName,
+              style: PetlyTextTheme.subtitle.copyWith(
+                fontSize: 14,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
           ],
         ),
       ),
